packages/fone_pipeline_analize_filler/pyproject.toml
[tool.poetry]
name = "fone-pipeline-analize-filler"
version = "0.1.0"
description = ""
authors = ["MikeTS, mike.tsvirkunov@gmail.com"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python=">=3.11,<3.13"
dotenv=">=0.9.9,<0.10.0"
logging=">=0.4.9.6,<0.5.0.0"
requests=">=2.32.5,<3.0.0"
beautifulsoup4=">=4.13.5,<5.0.0"
confluent-kafka=">=2.11.1,<3.0.0"
psycopg2-binary=">=2.9.10,<3.0.0"
optimum = "^1.27.0"
sentence-transformers = "^5.1.1"
onnxruntime = "^1.23.0"
onnx = "^1.19.0"
polars = "^1.33.1"
spacy = "^3.8.7"
catboost = "^1.2.8"
scikit-learn = "^1.7.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
